create contest keyboard button and contest type select keyboard

+++ bot/handlers/post_message_handlers.py
@@ -17,7 +17,7 @@ from bot.post_message.post_message_utils import is_post_message_valid
 from bot.keyboards.post_message_keyboards import InlinePostMessageStartConfirmKeyboard, InlinePostMessageMenuKeyboard, \
     InlinePostMessageExtraSettingsKeyboard, InlinePostMessageAcceptDeletingKeyboard, UnknownPostMessageType
 from bot.post_message.post_message_editors import edit_delay_date, edit_message, edit_button_text, edit_button_url, \
-    edit_media_files, send_post_message, PostActionType
+    edit_media_files, send_post_message, PostActionType, edit_winners_count, edit_contest_finish_date
 from bot.handlers.mailing_settings_handlers import send_post_messages
 from bot.post_message.post_message_decorators import check_callback_conflicts
 from bot.post_message.post_message_callback_handler import post_message_handler
@@ -42,6 +42,8 @@ async def post_message_menu_callback_handler(
     States.EDITING_POST_BUTTON_URL,
     States.EDITING_POST_MEDIA_FILES,
     States.EDITING_POST_DELAY_DATE,
+    States.EDITING_CONTEST_FINISH_DATE,
+    States.EDITING_CONTEST_WINNERS_COUNT,
 ))
 async def editing_post_message_handler(message: Message, state: FSMContext):
     state_data = await state.get_data()
@@ -52,6 +54,8 @@ async def editing_post_message_handler(message: Message, state: FSMContext):
             post_message_type = PostMessageType.MAILING
         case PostMessageType.CHANNEL_POST.value:
             post_message_type = PostMessageType.CHANNEL_POST
+        case PostMessageType.CONTEST.value:
+            post_message_type = PostMessageType.CONTEST
         case _:
             raise UnknownPostMessageType
 
@@ -66,6 +70,10 @@ async def editing_post_message_handler(message: Message, state: FSMContext):
             await edit_button_url(message, state, post_message_type)
         case States.EDITING_POST_MEDIA_FILES:
             await edit_media_files(message, state, post_message_type)
+        case States.EDITING_CONTEST_WINNERS_COUNT:
+            await edit_winners_count(message, state, post_message_type)
+        case States.EDITING_CONTEST_FINISH_DATE:
+            await edit_contest_finish_date(message, state, post_message_type)
 
 
 @post_message_router.callback_query(
@@ -86,21 +94,21 @@ async def post_message_extra_settings_callback_handler(
                 query,
                 post_message,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
         case callback_data.ActionEnum.NOTIFICATION_SOUND:
             await _notification_sound(
                 query,
                 post_message,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
         case callback_data.ActionEnum.BACK_TO_POST_MESSAGE_MENU:
             await _inline_back_to_post_message_menu(
                 query,
                 bot_id,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
 
 
@@ -122,14 +130,14 @@ async def post_message_confirm_start_callback_handler(
                 query,
                 post_message,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
         case callback_data.ActionEnum.BACK_TO_POST_MESSAGE_MENU:
             await _inline_back_to_post_message_menu(
                 query,
                 bot_id,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
 
 
@@ -151,14 +159,14 @@ async def post_message_accept_deleting_callback_handler(
                 query,
                 post_message,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
         case callback_data.ActionEnum.BACK_TO_POST_MESSAGE_MENU:
             await _inline_back_to_post_message_menu(
                 query,
                 bot_id,
                 post_message_type,
-                channel_id=callback_data.channel_id if post_message_type == PostMessageType.CHANNEL_POST else None
+                channel_id=callback_data.channel_id if post_message_type in (PostMessageType.CHANNEL_POST, PostMessageType.CONTEST) else None
             )
 
 
@@ -213,7 +221,7 @@ async def _start_confirm(
                     post_message.job_id = job_id
                     await post_message_db.update_post_message(post_message)
 
-            case PostMessageType.CHANNEL_POST:
+            case PostMessageType.CHANNEL_POST | PostMessageType.CONTEST:
                 channel_username = (await Bot(custom_bot.token).get_chat(channel_id)).username
 
                 await query.message.answer(
@@ -236,7 +244,12 @@ async def _start_confirm(
                         media_files,
                         PostActionType.RELEASE
                     )
-                    await post_message_db.delete_post_message(post_message.post_message_id)
+                    if post_message_type == PostMessageType.CONTEST:
+                        post_message.is_sent = True
+                        post_message.is_running = True
+                        await post_message_db.update_post_message(post_message)
+                    else:
+                        await post_message_db.delete_post_message(post_message.post_message_id)
 
                 else:
                     job_id = await _scheduler.add_scheduled_job(
@@ -278,7 +291,7 @@ async def _delete_post_message(
                 parse_mode=ParseMode.HTML
             )
 
-        case PostMessageType.CHANNEL_POST:
+        case PostMessageType.CHANNEL_POST | PostMessageType.CONTEST:
             channel_username = (await Bot(custom_bot.token).get_chat(channel_id)).username
             keyboard = await InlineChannelMenuKeyboard.get_keyboard(post_message.bot_id, channel_id)
 
@@ -354,7 +367,7 @@ async def _inline_back_to_post_message_menu(
     match post_message_type:
         case PostMessageType.MAILING:
             username = (await Bot(custom_bot_token).get_me()).username
-        case PostMessageType.CHANNEL_POST:
+        case PostMessageType.CHANNEL_POST | PostMessageType.CONTEST:
             username = (await Bot(custom_bot_token).get_chat(channel_id)).username
         case _:
             raise UnknownPostMessageType
